chore: rename package and directory names
