+++ app/src/main/java/agent/App.java
@@ -1,7 +1,4 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package org.example;
+package agent;
 
 public class App {
   public String getGreeting() {
