+++ app/src/test/java/agent/AppTest.java
@@ -1,7 +1,4 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package org.example;
+package agent;
 
 import static org.junit.jupiter.api.Assertions.*;
